+++ backend/blog/src/main/java/com/jk/blog/service/CategoryService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface CategoryService {
 
     CategoryDTO createCategory(CategoryDTO categoryDTO);
-    CategoryDTO updateCategory(CategoryDTO categoryDTO, Integer categoryId);
-    CategoryDTO getCategoryById(Integer categoryId);
+    CategoryDTO updateCategory(CategoryDTO categoryDTO, Long categoryId);
+    CategoryDTO getCategoryById(Long categoryId);
     List<CategoryDTO> getAllCategories();
-    void deleteCategory(Integer categoryId);
+    void deleteCategory(Long categoryId);
 }
